requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
pyyaml==6.0.2
aiosqlite==0.20.0
watchfiles==1.0.4
python-multipart==0.0.20
pybase64==1.4.1
